ejemplo_clase/requirements.txt
Django==3.2.2
requests==2.25.1
django-filter==2.4.0
djangorestframework==3.12.4
django-rest-auth==0.9.5
# Swagger:
uritemplate==3.0.1
pyyaml==5.4.1
# Drivers para trabajar con Postgres en linux:
psycopg2-binary>=2.8
# Graficos:
plotly==5.3.1
# Tareas asincrónicas:
celery==5.1
redis==3.3.11
# Google API:
google-api-python-client==2.19.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6